refactor: move args parser to args.js
